A6: split CustomFieldRepository into Reader + Writer behind a façade (#563 backlog)
First commit of the repositories read/write split. Extracts the custom-field
repository's two concerns into dedicated classes:

- CustomFieldReader — every SELECT / lookup / derived-read query plus the pure
  field-introspection helpers (choices, validation rules, validate_field_value).
- CustomFieldWriter — every INSERT / UPDATE / DELETE plus write-only helpers
  (key generation, uniqueness, sensitive-registry cache invalidation).

CustomFieldRepository stays the public entry point as a thin delegating façade,
so the ~38 call sites and the public constants (FIELD_TYPES, USER_META_KEY, …)
need no change. Reader and Writer share the same cache group so writes still
invalidate the entries reads populate. A tech-debt note on the façade points to
the future call-site migration (#563 B3).

Behavior-preserving — CustomFieldRepositoryTest + all reregistration dependents
green (513 tests); PHPStan L8 + WPCS clean.
